cheri: Fix compile for drivers/net/wireless/marvell/
Pointers in structures read from hardware are actually
addresses. Make them unsigned long to avoid CHERI related
compiler errors. Add casts and uintptr_t as necessary.

+++ drivers/net/wireless/marvell/libertas/cfg.c
@@ -540,7 +540,7 @@ static int lbs_cfg_set_mesh_channel(struct wiphy *wiphy,
 #define LBS_SCAN_RSSI_TO_MBM(rssi) \
 	((-(int)rssi + 3)*100)
 
-static int lbs_ret_scan(struct lbs_private *priv, unsigned long dummy,
+static int lbs_ret_scan(struct lbs_private *priv, uintptr_t dummy,
 	struct cmd_header *resp)
 {
 	struct cfg80211_bss *bss;
