+++ drivers/net/wireless/marvell/libertas_tf/libertas_tf.h
@@ -327,8 +327,8 @@ struct cmd_ctrl_node {
 	int result;
 	/* command response */
 	int (*callback)(struct lbtf_private *,
-			unsigned long, struct cmd_header *);
-	unsigned long callback_arg;
+			uintptr_t, struct cmd_header *);
+	uintptr_t callback_arg;
 	/* command data */
 	struct cmd_header *cmdbuf;
 	/* wait queue */
@@ -509,9 +509,9 @@ void lbtf_cmd_async(struct lbtf_private *priv, uint16_t command,
 
 int __lbtf_cmd(struct lbtf_private *priv, uint16_t command,
 	      struct cmd_header *in_cmd, int in_cmd_size,
-	      int (*callback)(struct lbtf_private *, unsigned long,
+	      int (*callback)(struct lbtf_private *, uintptr_t,
 			      struct cmd_header *),
-	      unsigned long callback_arg);
+	      uintptr_t callback_arg);
 
-int lbtf_cmd_copyback(struct lbtf_private *priv, unsigned long extra,
+int lbtf_cmd_copyback(struct lbtf_private *priv, uintptr_t extra,
 		     struct cmd_header *resp);
